+++ src/azure-cli/azure/cli/command_modules/network/_help.py
@@ -4424,6 +4424,41 @@ examples:
     text: az network vnet-gateway wait -g MyResourceGroup -n MyVnetGateway --created
 """
 
+helps['network vnet-gateway aad'] = """
+type: group
+short-summary: Manage AAD(Azure Active Directory) authentication of a virtual network gateway
+"""
+
+helps['network vnet-gateway aad assign'] = """
+type: command
+short-summary: Assign/Update AAD(Azure Active Directory) authentication to a virtual network gateway.
+examples:
+  - name: Assign AAD authentication to a virtual network gateway
+    text: |-
+        az network vnet-gateway aad assign \\
+        --resource-group MyResourceGroup \\
+        --gateway-name MyVnetGateway \\
+        --tenant MyAADTenantURI \\
+        --audience MyAADAudienceId \\
+        --issuer MyAADIssuerURI
+"""
+
+helps['network vnet-gateway aad show'] = """
+type: command
+short-summary: Show AAD(Azure Active Directory) authentication of a virtual network gateway
+examples:
+  - name: Show AAD information
+    text: az network vnet-gateway aad show --resource-group MyResourceGroup --gateway-name MyVnetGateway
+"""
+
+helps['network vnet-gateway aad remove'] = """
+type: command
+short-summary: Remove AAD(Azure Active Directory) authentication from a virtual network gateway
+examples:
+  - name: Remove AAD information
+    text: az network vnet-gateway aad remove --resource-group MyResourceGroup --gateway-name MyVnetGateway
+"""
+
 helps['network vpn-connection'] = """
 type: group
 short-summary: Manage VPN connections.